Add VP-Support for TimeRecords-export
